replace deprecated AESFastEngine with AESEngine
from the bouncycastle release notes:
CVE-2016-1000339:
AESFastEngine has a side channel leak if table accesses can be observed.
The use of lookup large static lookup tables in AESFastEngine means that
where data accesses by the CPU can be observed,
it is possible to gain information about the key used to initialize the cipher.
We now recommend not using AESFastEngine where this might be a concern.
The BC provider is now using AESEngine by default.